ApiReferenceService
This service is responsible to check the validity of the supported APIs passed to the application.
Currently only one API (Imdb - using Omdb unofficial API) is supported. The provision of flexibility
is provided where support for more APIs can further be added. The design takes care of that. Further
APIs can be added in the Application Context and implementing class (Currently under development)
can be added. This would add support to new APIs.