MovieParserService
This is an interface that is responsible to provide the skeleton for connecting to the various Movie APIs
One implementation of the same is provided here with ImdbParserServiceImpl, which uses the IMDB
Rest API (Unofficial IMDB API used - OMDB). More such implementations can be added (using factory
method implementation) to add support for more and more APIs.
The test case for the MovieParserService mocks the calling of the service to test the functionality in
isolation, i.e. without the web service being called